Refactor out BaseWindowsMixedRealitySource
